# backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "CompetitionPublic": {
        "properties": {
          "competition_gender": {
            "maxLength": 20,
            "title": "Competition Gender",
            "type": "string"
          },
          "competition_international": {
            "default": false,
            "title": "Competition International",
            "type": "boolean"
          },
          "competition_name": {
            "maxLength": 255,
            "title": "Competition Name",
            "type": "string"
          },
          "competition_youth": {
            "default": false,
            "title": "Competition Youth",
            "type": "boolean"
          },
          "country_name": {
            "maxLength": 100,
            "title": "Country Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "match_available": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Available"
          },
          "match_available_360": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Available 360"
          },
          "match_count": {
            "default": 0,
            "title": "Match Count",
            "type": "integer"
          },
          "match_updated": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Updated"
          },
          "match_updated_360": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Updated 360"
          },
          "season_id": {
            "title": "Season Id",
            "type": "integer"
          },
          "season_name": {
            "maxLength": 100,
            "title": "Season Name",
            "type": "string"
          },
          "statsbomb_id": {
            "title": "Statsbomb Id",
            "type": "integer"
          }
        },
        "required": [
          "statsbomb_id",
          "season_id",
          "id",
          "country_name",
          "competition_name",
          "competition_gender",
          "season_name"
        ],
        "title": "CompetitionPublic",
        "type": "object"
      },
      "CompetitionsPublic": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/CompetitionPublic"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data",
          "count"
        ],
        "title": "CompetitionsPublic",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestResult": {
        "properties": {
          "imported_competitions": {
            "title": "Imported Competitions",
            "type": "integer"
          },
          "imported_matches": {
            "title": "Imported Matches",
            "type": "integer"
          }
        },
        "required": [
          "imported_competitions",
          "imported_matches"
        ],
        "title": "IngestResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "<PROJECT_NAME>",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/soccer/competitions/": {
      "get": {
        "operationId": "readCompetitions",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "has_matches",
            "required": false,
            "schema": {
              "default": false,
              "title": "Has Matches",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "has_events",
            "required": false,
            "schema": {
              "default": false,
              "title": "Has Events",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionsPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Competitions",
        "tags": [
          "soccer"
        ]
      }
    },
...
    "/api/v1/soccer/competitions/ingest": {
      "post": {
        "operationId": "ingestSoccerData",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Ingest Soccer Data",
        "tags": [
          "soccer"
        ]
      }
    },
...
  }
}